+++ src/test/suite/index.ts
@@ -2,6 +2,15 @@ import * as glob from 'glob';
 import * as Mocha from 'mocha';
 import * as path from 'path';
 
+// 2023-01-15 removed mocha and @types/mocha packages so these integration testd will no longer run.
+// Why? Because of conflicts with jest (reference: https://stackoverflow.com/a/64202454)
+// that manifested as a series of errors in the Github check a la:
+//   Error: node_modules/@types/jest/index.d.ts(34,13): error TS2403: Subsequent variable declarations must have the same type.
+//            Variable 'beforeEach' must be of type 'HookFunction', but here has type 'Lifecycle'.
+// Can switch this over to jest down the road, per Soloydenko's article
+// (https://medium.com/@soloydenko/end-to-end-testing-vs-code-extensions-via-jest-828e5edfeb75)
+// Also see https://stackoverflow.com/q/49615315
+
 export function run(): Promise<void> {
 	// Create the mocha test
 	const mocha = new Mocha({