+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to the "vscode-styra" extension will be documented in this f
 
 ## [Unreleased]
 
-N/A
+### Added
+
+- Unit test infrastructure added
+- Unit test PR gating check added
+- API seam for VSCode, allowing for ease of adding other IDEs down the road
 
 ## [0.0.1] - 2023-01-10